add authorization token to query

+++ frontend/src/components/NoteDetails.js
@@ -1,4 +1,5 @@
 import { useNoteContext } from "../hooks/useNoteContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 // date
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
@@ -6,10 +7,19 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 const NoteDetails = ({ note }) => {
 
     const { dispatch } = useNoteContext();
+    const { user } = useAuthContext();
 
     const handleClick = async () => {
+
+        if (!user) {
+            return;
+        }
+
         const response = await fetch('http://localhost:4000/api/notes/' + note._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         });
 
         const json = await response.json();
